Removed old/unused code
