+++ eoflow/base/__init__.py
@@ -2,5 +2,3 @@ from .configuration import Configurable
 from .base_task import BaseTask
 from .base_input import BaseInput
 from .base_model import BaseModel, ModelMode
-from .base_train import BaseTrain
-from .base_predict import BasePredict
