+++ eoflow/models/__init__.py
@@ -1,5 +1,4 @@
 from .example_model import ExampleModel
-from .template_model import TemplateModel
 from .fcn_model import FCNModel
 from .tfcn_model import TFCNModel
 from .rfcn_model import RFCNModel
