--- eoflow/predictors/__init__.py
@@ -1,1 +1,0 @@
-from .fcn_predictor import TFCNPredict
