--- eoflow/trainers/__init__.py
@@ -1,3 +1,0 @@
-from .example_trainer import ExampleTrainer
-from .template_trainer import TemplateTrainer
-from .fcn_trainer import FCNTrainer
