+++ eoflow/utils/__init__.py
@@ -1,4 +1,2 @@
-from .config import get_config_from_json, process_config, parse_classname
-from .dirs import create_dirs
+from .utils import create_dirs, parse_classname
 from .logger import Logger
-from .utils import get_args
