--- eoflow/utils/dirs.py
@@ -1,17 +1,0 @@
-import os
-
-
-def create_dirs(dirs):
-    """
-    dirs - a list of directories to create if these directories are not found
-    :param dirs:
-    :return exit_code: 0:success -1:failed
-    """
-    try:
-        for dir_ in dirs:
-            if not os.path.exists(dir_):
-                os.makedirs(dir_)
-        return 0
-    except Exception as err:
-        print("Creating directories error: {0}".format(err))
-        exit(-1)
